[Refactor]添加注释，大概还原了一些删掉的部分
