+++ lib/pages/CompanionPage.dart
@@ -4,6 +4,9 @@ import 'package:intellij_tourism_designer/constants/Themes.dart';
 import 'package:intellij_tourism_designer/constants/location.dart';
 import 'package:intellij_tourism_designer/widgets/DetailView.dart';
 import 'package:intellij_tourism_designer/widgets/MapView.dart';
+
+//大概是人已经在路上的模块，还没想好叫什么名字
+
 class CompanionPage extends StatefulWidget {
   const CompanionPage({super.key});
 
@@ -16,7 +19,7 @@ class _CompanionPageState extends State<CompanionPage> {
   Widget build(BuildContext context) {
     return Stack(
       children:[
-        MapWidget(),
+        DemoMap(),
         Positioned(
           right:20,top:20,width:340,height:100,
           child: Card(
