+++ lib/pages/EditTracePage.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+//import 'package:intellij_tourism_designer/widgets/Drawer.dart';
+class EditTracePage extends StatefulWidget {
+  const EditTracePage({super.key});
+
+  @override
+  State<EditTracePage> createState() => _EditTracePageState();
+}
+
+class _EditTracePageState extends State<EditTracePage> {
+  Widget build(BuildContext context) {
+    return Scaffold(
+      //drawer: const MenuDrawer(EditTracePage.route),
+    );
+  }
+}
