+++ lib/pages/PickUpPage.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+class PickUpPage extends StatelessWidget {
+  const PickUpPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      
+    );
+  }
+
+}
