+++ lib/widgets/ItiFeature.dart
@@ -3,6 +3,8 @@ import 'package:easy_date_timeline/easy_date_timeline.dart';
 import 'package:intellij_tourism_designer/constants/Themes.dart';
 import 'package:intellij_tourism_designer/constants/location.dart';
 
+//自动规划行程的页面，主要时一些个性设置
+
 class ItiFeature extends StatefulWidget {
   const ItiFeature({super.key});
