+++ lib/widgets/MapView.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:intellij_tourism_designer/constants/Themes.dart';
 import 'package:intellij_tourism_designer/constants/location.dart';
@@ -7,15 +8,16 @@ import 'package:latlong2/latlong.dart';
 import 'package:intellij_tourism_designer/widgets/TileProviders.dart';
 import 'package:intellij_tourism_designer/constants/Markers.dart';
 
-class MapWidget extends StatefulWidget {
-  const MapWidget({super.key});
+class DemoMap extends StatefulWidget {
+  const DemoMap({super.key});
   @override
-  State<MapWidget> createState() => _PraticeState();
+  State<DemoMap> createState() => _PraticeState();
 }
-class _PraticeState extends State<MapWidget> {
+class _PraticeState extends State<DemoMap> {
   late final MapController _mapController;
   TileMap map = TileMap(MapServiceProvider.osm); // The map here is a state.
   late var _markers = MarkerList.normalBookmark;
+  bool _isDetail = false;
 
   void _changeMapProviderProvider(MapServiceProvider provider) {
     setState(() {
@@ -47,66 +49,115 @@ class _PraticeState extends State<MapWidget> {
   @override
   Widget build(BuildContext context) {
     return LayoutBuilder(builder:  (context, constraints) {
-          return Stack(
-            alignment: Alignment.topRight,
-              children: [
-                 FlutterMap(
+      return Column(
+        children:[
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8),
+            child: Row(
+              children: <Widget>[
+                MaterialButton(
+                  onPressed: () {setState(() {
+                    if (_isDetail) {
+                      _changeBookmarkLevel(MarkerLevel.normal);
+                    }
+                    else{
+                      _changeBookmarkLevel(MarkerLevel.detailed);
+                    }
+                    _isDetail = !_isDetail;
+                    });
+                  },
+                  child: _isDetail? const Text('隐藏未来城校园信息') : const Text('显示未来城校园信息'),
+                ),
+                MaterialButton(
+                  onPressed: () => _mapController.move(Location.nanwangshan, 15),
+                  child: const Text('南望山校区'),
+                ),
+                MaterialButton(
+                  onPressed: () => _mapController.move(Location.weilaicheng, 15),
+                  child: const Text('未来城校区'),
+                ),
+                MaterialButton(
+                  onPressed: () => _mapController.move(Location.home, 10),
+                  child: const Text('我的家乡'),
+                ),
+                MaterialButton(
+                  onPressed: () => _mapController.move(Location.wuhan, 7),
+                  child: const Text('武汉市'),
+                ),
+              ]
+            )
+          ),
+          Flexible(
+            fit: FlexFit.tight,
+            child: Stack(
+              alignment: Alignment.topRight,
+                children: [
+                  FlutterMap(
                     mapController: _mapController,
                     options: MapOptions(
+                      //initialCenter: Location.nanwangshan,
                       initialCenter: Location.wuhan,
                       initialZoom: 17,
                       maxZoom: 19,
                       minZoom: 10,
                       cameraConstraint: CameraConstraint.contain(
-                        bounds: LatLngBounds( // unlimit the map range 
+                        bounds: LatLngBounds( // unlimit the map range
                           const LatLng(-90, -180),
                           const LatLng(90,180),
+                          ),
                         ),
                       ),
+                      children: [
+                        map.map,
+                        MarkerLayer(markers: _markers),
+                        RichAttributionWidget(
+                          popupInitialDisplayDuration: const Duration(seconds: 5),
+                          animationConfig: const ScaleRAWA(),
+                          showFlutterMapAttribution: false,
+                          attributions: [
+                            map.info,
+                            const TextSourceAttribution(
+                              '这是一个使用Flutter web呈现地图服务的demo,用于tcb提交WebGIS作业.'
+                              '选择Flutter的原因是Flutter可以非常方便地做跨平台分发.'
+                              '因此该网页使用Dart脚本而非JavaScript脚本.',
+                              prependCopyright: false,)
+                          ],
+                        ),
+                      ],
                     ),
-                    children: [
-                      map.map,
-                      MarkerLayer(markers: _markers),
-                      RichAttributionWidget(
-                              popupInitialDisplayDuration: const Duration(seconds: 5),
-                              animationConfig: const ScaleRAWA(),
-                              showFlutterMapAttribution: false,
-                              attributions: [
-                                map.info
-                              ],
-                            ),
-                          ],    
-                    ),
-                      Column(
-                          children: <Widget>[
-                            TextButton(
-                              onPressed: () {
-                                _changeMapProviderProvider(MapServiceProvider.osm);
-                                _changeBookmarkLevel(MarkerLevel.normal);
-                              },
-                              style:AppButton.button3,
-                              child: const Text('OSM'),
-                            ),
-                            TextButton(
-                              onPressed: () {
-                                _changeMapProviderProvider(MapServiceProvider.tianditu);
-                                _changeBookmarkLevel(MarkerLevel.normal);
-                              },
-                              style:AppButton.button3,
-                              child: const Text('天地图'),
-                            ),
-                            TextButton(
-                              onPressed: () {
-                                _changeMapProviderProvider(MapServiceProvider.gaode);
-                                _changeBookmarkLevel(MarkerLevel.none);
-                              },
-                              style:AppButton.button3,
-                              child: const Text('高德'),
-                            ),
-                          ]
-                      )
-              ],
-            );
+                    Column(
+                      children: <Widget>[
+                        TextButton(
+                          onPressed: () {
+                            _changeMapProviderProvider(MapServiceProvider.osm);
+                            _changeBookmarkLevel(MarkerLevel.normal);
+                            },
+                          style:AppButton.button3,
+                          child: const Text('OSM'),
+                        ),
+                        TextButton(
+                          onPressed: () {
+                            _changeMapProviderProvider(MapServiceProvider.tianditu);
+                            _changeBookmarkLevel(MarkerLevel.normal);
+                          },
+                          style:AppButton.button3,
+                          child: const Text('天地图'),
+                        ),
+                        TextButton(
+                          onPressed: () {
+                            _changeMapProviderProvider(MapServiceProvider.gaode);
+                            _changeBookmarkLevel(MarkerLevel.none);
+                          },
+                          style:AppButton.button3,
+                          child: const Text('高德'),
+                        ),
+                      ]
+                    )
+                  ],
+                ),
+          )
+            ]
+          );
         }
     );
   }
